requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Model serialization
joblib>=1.3.0

# Configuration
pyyaml>=6.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Web Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0
requests>=2.31.0  # For Ollama API calls

# Optional: Advanced NLP (uncomment if needed)
# sentence-transformers>=2.2.2
# transformers>=4.35.0
# torch>=2.1.0
